newport/gsc: add support for new ADC values returned from GSC v51+
The GSC used on Ventana and Laguna pre-scales ADC values based on
hard-coded per-board data. The Newport GSC adopted this practice as well
but starting with firmware v51 it will instead return raw ADC values
for the voltage rails.

This has the following benefits:
 - greatly reduces possibility for NAK's
 - improves the rate at which ADC's can be read. Before they were read
   and cooked (scaled/adjusted) only once a second, now they are continually
   read (each value should be updated once per ~20us depending on board)

The only downside of this is that now the ADC's have to be scaled by the
driver or application interpreting them. We are using the device tree to
define the reference voltage, ADC resolution, and voltage divider pre-scaler
for this thus the equation for conversion is:
  Vo = Vadc * (Vref / resolution) * (R2+R1/R2)

Signed-off-by: Tim Harvey <tharvey@gateworks.com>