newport: remove microSD from dts if not present on boot
The microSD on Newport should never be ejected or inserted at runtime
as it can glitch the data bus shared with eMMC.

Additionally having an unpopulated microSD slot causes the MMC driver to
continually poll for presence of a microSD (because of the lack of a
dedicated card detect signal) and this can cause I/O to block for
several seconds periodically when an eMMC filesystem is mounted.

This patch will remove the microSD socket from the device-tree if a
microSD card was not present at boot.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>