newport: add SFP support
Detect and configure VSC8574 QSGMII PHY ports for SFP modules.

This requires the SFP module be inserted on boot and not removed
or replaced while the board is running.

This is working for 1000BASE-X (Gig Fiber) modules but is currently
not working for 10/100/1000BASE-T (copper) modules or 100BASE-FX
(100mbps Fiber) modules.

Tested with:
- ADDON MGBSX1-AO A1 ADDSX20795 1000Base-SX Fiber (850nm multi-mode fiber with LC connector)
- FINISAR CORP. FTRJ-8519-7D P4505PR 1000Base-SX Fiber (850nm multi-mode with LC connector)

Signed-off-by: Tim Harvey <tharvey@gateworks.com>