boards: refactor dts files to define unique DRAM config in base model
No meaningful changes in resulting DTB's with the exception of the DRAM
SPD's of gwxx0 vs gwxx4 models.

The Newport BDK will dynamically adjust the DRAM SPD based on DRAM
configuration detected in board info EEPROM. However specifying it
in the model specific dts instead of the shared dtsi makes it easier
to use the Cavium SDK BDK by just dropping in the Newport DTB's.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>