libdram: Use multiple patterns for read leveling
The default pattern in read leveling is 0x55, some other cores
utilize 0x33 for the read leveling pattern (MR3, Page 0 = 0x1).

There are a total of 4 patterns available in the MR3, Page 0 register
that can be utilized during the read leveling sequence as follows:
0 = 0x55
1 = 0x33
2 = 0x0f
3 = 0x00

Pattern 3 is not a pattern of 0's and 1's thus is not a good option
as a pattern, so for now just don't use it. Instead loop through the
3 other patterns 3 times each when sampling the read level to find
the best value.

0x55 consistantly shows us poor results on newport boards
by having results with 0 errors but rlevel delays of 11 which we known
are too high so we will only use 0x33 and 0x0f.

Signed-off-by: Chris Lang <clang@gateworks.com>