bdk-mmc: reduce MMC bus speed for booting from microSD
Revert the MMC clock speed bump and set it back to 10MHz for
compatibility with SD cards in legacy mode. This fixes booting from
microSD and testing has shown it has a negligable effect on boot speed
and still gets to U-Boot in 6 seconds when booting from eMMC with or
with an SD present.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>