bdk-mmc: Support working with more than one mmc cmd
The cn81xx device has a single 8-bit mmc bus that is shared between
3 command (chip select) signals. This allows communication to 3
distinct chips, however, the data bus is still shared between the
3 devices. In order to properly communicate with each device, we
need to keep track of the last device that we were communicating with
and switch off the last device if we are going to communicate with
a new device on a different command signal.

Additionally, cleanup mmc_initialize and mimic the proven setup of
u-boot and linux.

This commit was inspired by u-boot/linux code.

Signed-off-by: Chris Lang <clang@gateworks.com>