newport: set gpio12 to UART2 RTS for rs485
When using RS485 the RTS signal is used as the tranceiver TX enable.
Setup the pinsel to change gpio12 to UART2 RTS instead of a GPIO.

Note this assumes you are using a Linux kernel uart driver
that supports the TIOCSRS485 ioctl.

For CN803x with "arm,pl011" uart the driver (drivers/tty/serial/amba-pl011.c)
added support for TIOCSRS485 ioctl in Linux 5.15 and this was backported
to the Gateworks 5.10 kernel.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>