newport: apply RGM drive strength fix for GPY111 per board model
commit 97b95939e164 ("newport: store PHY config in dt for U-Boot/Linux
PHY drivers") adjusted CN803x RGM drive strength to the maxiumum to
compensate for packet corruption errors with the GPY111 PHY. Testing has
shown that this causes excessive EMI at 125MHz and that it is only
necessary when the GPY111 internal DC-to-DC is used for the 1.0V rail.

Only apply the drive strength changes for the board models that have the
GPY111 PHY where its internal DC-to-DC is used for the 1.0V rail.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>